Also move `objectSpread2` (to show an helper with deps)
